/test: consolidate governor tests, fix the TUI's own no-checks conflation
Test suite:
- Folded the new PrsContextGithubPollGovernor.test.tsx into PrsContext.test.tsx
  (31 tests). The folder was over its per-folder budget; the two files that
  remain each have a structural reason — PrsContextWarmCache needs module-load
  isolation (MODE=production before a dynamic import) and githubPollGovernor is
  a pure node-shaped state machine colocated with its module.

TUI parity — a real defect, one layer below the one this branch fixes:
- `/pr` dispatch catches a rejected read into `{ error }`, but the right-pane
  formatters treated that shape as "no rows" and printed "No PR checks." for a
  GitHub outage. That is exactly the failure-reads-as-empty conflation
  `prService.getChecks` stopped making when it started rejecting instead of
  returning `[]`, reproduced at the render step — and "no checks ran" invites a
  merge. `formatPrChecks`, `formatPrComments`, and `formatPrReview` now name the
  failed sources, and a source that failed reports no count rather than "0
  reviews". Partial success still renders whatever answered.

Docs (four inaccuracies in prose written across the review rounds):
- The refresh() sweep needs BOTH conditions (every candidate failed AND the
  reason is GitHub-wide), not just the second.
- "lengthens the period rather than skipping ticks" is true of the detail pane's
  loops but not the provider's 60s poll, which deliberately skips.
- The ladder base is 60s only for confirmed-broken kinds; network/unknown climb
  from 30s and rate_limited goes straight to the ceiling. Re-derivation happens
  only when the kind changed — the livelock guard was undocumented.
- The source-file-map row still said the budget scans "a PR-read resource",
  contradicting the unknown-bucket fix documented below it.
- Added the missing surfaces: the sync remote-command registry, ARCHITECTURE's
  IPC channel list, and the transport-failure recording in both owners.

Mobile: no changes required. iOS already reads `unavailableParts` and retains
the last-good checks rather than adopting an empty array, with existing
coverage; it never invokes a github.* remote command and never polls GitHub.
