+++ apps/ade-cli/src/tuiClient/rightPaneFormatters.ts
@@ -71,6 +71,23 @@ function unwrapPrValue(value: unknown): { root: JsonRecord; pr: JsonRecord } | n
   };
 }
 
+/**
+ * The message from a PR read that failed, or null when the value is real data.
+ *
+ * The `/pr` dispatch catches a rejected read into `{ error }` rather than
+ * blowing the pane away, so every PR formatter has to tell "GitHub refused"
+ * apart from "there is nothing to show". Without this the two collapse: an
+ * `{ error }` carries no rows, so the formatters below printed "No PR checks."
+ * for an outage — the exact conflation `prService.getChecks` stopped making
+ * when it started rejecting instead of returning `[]`. A reader cannot act on
+ * a lie that reads as a fact, and "no checks ran" invites a merge.
+ */
+function prReadFailure(value: unknown): string | null {
+  const root = unwrapStructured(value);
+  if (!isRecord(root)) return null;
+  return asString(root.error);
+}
+
 function firstRecordArray(value: unknown, keys: string[]): JsonRecord[] {
   const root = unwrapStructured(value);
   if (Array.isArray(root)) return root.filter(isRecord);
@@ -354,6 +371,8 @@ export function formatPrMergeState(value: unknown): string {
 }
 
 export function formatPrChecks(value: unknown): string {
+  const failure = prReadFailure(value);
+  if (failure) return `PR checks · could not be read — ${failure}`;
   const root = unwrapStructured(value);
   const rollup = isRecord(root) ? pickString(root, ["checksStatus"]) : null;
   const reason = isRecord(root) ? pickString(root, ["checksReason"]) : null;
@@ -404,8 +423,35 @@ export function formatPrReview(value: unknown): string {
   const reviews = firstRecordArray(root, ["reviews"]);
   const threads = firstRecordArray(root, ["reviewThreads", "threads"]);
   const comments = firstRecordArray(root, ["comments", "issueComments"]);
+  // Three independent reads land here, each caught into its own `{ error }`,
+  // so a partial outage is normal: name the sources that failed rather than
+  // reporting their absence as "0 reviews".
+  const failures: Array<[label: string, message: string]> = [];
+  for (const [label, part] of [
+    ["reviews", isRecord(root) ? root.reviews : null],
+    ["threads", isRecord(root) ? (root.reviewThreads ?? root.threads) : null],
+    ["comments", isRecord(root) ? (root.comments ?? root.issueComments) : null],
+  ] satisfies Array<[string, unknown]>) {
+    const message = prReadFailure(part);
+    if (message) failures.push([label, message]);
+  }
+  const rootFailure = prReadFailure(root);
+  if (rootFailure) return `PR review · could not be read — ${rootFailure}`;
+  if (failures.length > 0 && !reviews.length && !threads.length && !comments.length) {
+    return [
+      "PR review · could not be read",
+      ...failures.map(([label, message]) => `  ✗ ${label}: ${message}`),
+    ].join("\n");
+  }
+  // A source that failed has no count to report — printing `0 reviews` would
+  // state as fact the very thing the read could not establish.
+  const failed = new Set(failures.map(([label]) => label));
+  const headerCount = (label: string, noun: string, count: number): string =>
+    failed.has(label) ? `${noun}s unavailable` : formatCount(noun, count);
   const lines = [
-    `PR review · ${formatCount("review", reviews.length)} · ${formatCount("thread", threads.length)} · ${formatCount("comment", comments.length)}`,
+    `PR review · ${headerCount("reviews", "review", reviews.length)}`
+    + ` · ${headerCount("threads", "thread", threads.length)}`
+    + ` · ${headerCount("comments", "comment", comments.length)}`,
   ];
   if (reviews.length) {
     lines.push("", "Reviews");
@@ -431,11 +477,18 @@ export function formatPrReview(value: unknown): string {
       lines.push(`- ${commentPreview(comment)}`);
     }
   }
-  if (!reviews.length && !threads.length && !comments.length) lines.push("", "No PR reviews or comments.");
+  if (failures.length > 0) {
+    lines.push("", "Could not be read");
+    for (const [label, message] of failures) lines.push(`  ✗ ${label}: ${message}`);
+  } else if (!reviews.length && !threads.length && !comments.length) {
+    lines.push("", "No PR reviews or comments.");
+  }
   return lines.join("\n");
 }
 
 export function formatPrComments(value: unknown): string {
+  const failure = prReadFailure(value);
+  if (failure) return `PR comments · could not be read — ${failure}`;
   const root = unwrapStructured(value);
   const summary = isRecord(root) && isRecord(root.summary) ? root.summary : null;
   const threads = firstRecordArray(root, ["reviewThreads", "threads"]);
